refactor: Chronology field array file structure
Mored chronology field intop the chronology field array file
