+++ src/components/RulesetFormSections/ChronologyFieldArray/ChronologyField/ChronologyField.test.js
@@ -7,11 +7,11 @@ import {
 
 import ChronologyField from './ChronologyField';
 
-import { translationsProperties } from '../../../../test/helpers';
-import { refdata as mockRefdata } from '../../../../test/resources';
+import { translationsProperties } from '../../../../../test/helpers';
+import { refdata as mockRefdata } from '../../../../../test/resources';
 
-jest.mock('../../utils', () => ({
-  ...jest.requireActual('../../utils'),
+jest.mock('../../../utils', () => ({
+  ...jest.requireActual('../../../utils'),
   useSerialsManagementRefdata: () => mockRefdata,
 }));
