+++ src/components/RulesetFormSections/index.js
@@ -5,7 +5,6 @@ export { default as IssuePublicationFieldArray } from './IssuePublicationFieldAr
 export { default as IssuePublicationField } from './IssuePublicationField';
 export { default as OmissionField } from './OmissionField';
 export { default as OmissionFieldArray } from './OmissionFieldArray';
-export { default as CombinationField } from './CombinationField';
 export { default as CombinationFieldArray } from './CombinationFieldArray';
 export { default as ChronologyFieldArray } from './ChronologyFieldArray';
 export { default as EnumerationFieldArray } from './EnumerationFieldArray';
